requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.2.2
pytest-cov==5.0.0
ruff==0.4.9